refactor(landing): split page.tsx into section components + fix primary glow bug
PR 1 of landing rebuild. Pure structural work — no copy or layout changes
visible to users, aside from the primary glow effects finally rendering.

- Split 681-line src/app/page.tsx into six section components under
  src/components/landing/ (landing-nav, hero, features, comparison,
  deploy, final-cta). page.tsx is now a thin composition with the
  session redirect.
- Fix the broken rgba(var(--primary), ...) idiom used in six sites.
  --primary is oklch(...), so rgba() silently dropped every primary
  glow, shadow, and radial gradient on the page. Switched to
  color-mix(in oklch, var(--primary) N%, transparent).
- Drop the opacity-fading h1 gradient (foreground→foreground/50) that
  half-erased the headline.
- Remove dead classes: perspective-1000 (no 3D transform child) and
  mask-image-linear-gradient-to-r (not a valid Tailwind utility).

See docs/plans/2026-04-21-landing-page-rebuild-design.md for the full
rebuild design. PRs 2-5 to follow.
